# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holodeck Router UI API",
    "description": "REST API for managing VCF 9 Holodeck deployments, Day 2 operations, reservations, commands, user management, and email notifications.",
    "version": "1.2.0"
  },
  "servers": [
    { "url": "/", "description": "Current instance" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "holodeck-token",
        "description": "JWT token set as httpOnly cookie after login"
      }
    },
    "schemas": {
...
      "BackgroundJob": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "mode": { "type": "string" },
          "status": { "type": "string", "enum": ["running", "completed", "failed"] },
          "sessionName": { "type": "string" },
          "logFile": { "type": "string", "nullable": true, "description": "Path to saved tmux output log (set on completion)" },
          "startedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time", "nullable": true },
          "user": {
            "type": "object",
            "properties": {
              "displayName": { "type": "string" },
              "username": { "type": "string" }
            }
          }
        }
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/deployments": {
      "get": {
        "tags": ["Deployments"],
        "summary": "List all deployment jobs",
        "responses": {
          "200": {
            "description": "Job list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/BackgroundJob" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Deployments"],
        "summary": "Start a new deployment",
        "description": "Launches a Holodeck VCF deployment. labadmin+ with active reservation, or superadmin. Runs Import-HoloDeckConfig to load the selected config, then New-HoloDeckInstance to deploy.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mode", "instanceId"],
                "properties": {
                  "mode": { "type": "string", "enum": ["vvf", "management", "fullstack", "dualsite"] },
                  "instanceId": { "type": "string", "pattern": "^[a-zA-Z0-9]+$", "maxLength": 8, "description": "Alphanumeric only, max 8 characters" },
                  "configId": { "type": "string", "description": "HoloDeck config ID — deployment settings resolved from its cached JSON" },
                  "version": { "type": "string", "description": "VCF version override" },
                  "site": { "type": "string", "enum": ["a", "b"], "default": "a" },
                  "vsanMode": { "type": "string", "enum": ["ESA", "OSA"] },
                  "depotType": { "type": "string", "enum": ["Online", "Offline"] },
                  "dnsDomain": { "type": "string" },
                  "logLevel": { "type": "string", "enum": ["INFO", "DEBUG", "SUCCESS", "WARN", "ERROR"] },
                  "nsxEdgeClusterMgmtDomain": { "type": "boolean" },
                  "nsxEdgeClusterWkldDomain": { "type": "boolean" },
                  "deployVcfAutomation": { "type": "boolean" },
                  "deploySupervisorMgmtDomain": { "type": "boolean" },
                  "deploySupervisorWldDomain": { "type": "boolean" },
                  "workloadDomainType": { "type": "string", "enum": ["SharedSSO", "IsolatedSSO"] },
                  "provisionOnly": { "type": "boolean" },
                  "cidr": { "type": "array", "items": { "type": "string" } },
                  "vlanRangeStart": { "type": "array", "items": { "type": "integer" } },
                  "siteBCidr": { "type": "array", "items": { "type": "string" } },
                  "siteBVlanRangeStart": { "type": "array", "items": { "type": "integer" } },
                  "datastoreName": { "type": "string", "description": "Datastore name for nested VM placement (DeveloperMode)" },
                  "trunkPortGroupName": { "type": "string", "description": "Trunk port group / virtual network name for Site A (DeveloperMode)" },
                  "trunkPortGroupNameSiteB": { "type": "string", "description": "Trunk port group for Site B in dual-site deployments (DeveloperMode)" },
                  "clusterName": { "type": "string", "description": "vSphere cluster name (DeveloperMode)" },
                  "dcName": { "type": "string", "description": "vSphere datacenter name (DeveloperMode)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deployment started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "jobId": { "type": "string" },
                    "sessionName": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid mode or instanceId" },
          "403": { "description": "Insufficient role or no reservation" }
        }
      },
      "delete": {
        "tags": ["Deployments"],
        "summary": "Clear completed job history",
        "description": "Deletes all non-running jobs and cleans up tmux sessions. labadmin+.",
        "responses": {
          "200": {
            "description": "History cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": { "type": "integer" }
                  }
                }
              }
            }
          },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
...
  }
}